[TASK] Only modify lastTypoLinkUrl when EXT:headless is really used
When including the static TypoScript Template a simple flag in TypoScript Setup
will be added to check if ContentObjectRenderer->typolink()'s `lastTypoLinkUrl`
should be modified or not.
This could be useful when you have a TYPO3 Installation with multiple TypoScript
Templates and not every one uses EXT:headless, so the default typolink behavior
will be retained.

Resolves #113