Extended Dataprocessors (#407)
* Update FlexFormProcessor.php

Added the possibility to process data, which is in $processedData[$fieldName] instead of $processedData['data'][$fieldName]. 
We needed this to process flexform data in a result of the DatabaseQueryProcessor.

* Update DatabaseQueryProcessor.php

This makes it possible, to configure only one field of the result.
We have a link field, which wasn't parsed as typolink. But we didn't want to define every single field of the query. So we added the key "overrideFields".